+++ packages/react-router-fs-routes/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@react-router/fs-routes`
 
+## 7.9.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@react-router/dev@7.9.3`
+
 ## 7.9.2
 
 ### Patch Changes